+++ R/add_limits_extrapolation.R
@@ -1,4 +1,4 @@
-#' Add a control to a BiodiversityModel object to control extrapolation
+#' Add a control to a BiodiversityModel object to limit extrapolation
 #'
 #' @description One of the main aims of species distribution models (SDMs) is to project
 #' in space and time. For projections a common issue is extrapolation as - unconstrained -
@@ -7,7 +7,7 @@
 #' be related to an insufficient quantification of the niche (e.g. niche truncation
 #' by considering only a subset of observations within the actual distribution),
 #' in other cases there can also be general barriers or constraints that limit
-#' any projections (e.g. islands). This control method adds some of those options
+#' any projections (e.g. islands). This limit method adds some of those options
 #' to a model distribution object. Currently supported methods are:
 #'
 #' [*] \code{"zones"} - This is a wrapper to allow the addition of zones to a
@@ -23,10 +23,14 @@
 #' following Mesgaran et al. (2014). This method is also available in the [similarity()]
 #' function.
 #'
+#' [*] \code{"mess"} - Constraints the predictions using the
+#' Multivariate Environmental Similarity Surfaces (MESS) following Mesgaran et al. (2014).
+#' This method is also available in the [similarity()] function.
+#'
 #' [*] \code{"shape"} - This is an implementation of the 'shape' method introduced
 #' by Velazco et al. (2023). Through a user defined threshold it effectively limits
 #' model extrapolation so that no projections are made beyond the extent judged as
-#' defensible and informed by the training observations.
+#' defensible and informed by the training observations. **Not yet implemented!**
 #'
 #' See also details for further explanations.
 #'
@@ -35,7 +39,7 @@
 #' surface when intersected with input data (Default: \code{NULL}).
 #' @param method A [`character`] vector describing the method used for controlling
 #' extrapolation. Available options are \code{"zones"}, \code{"mcp"} (Default),
-#' or \code{"nt2"} or \code{"shape"}.
+#' or \code{"nt2"}, \code{"mess"} or \code{"shape"}.
 #' @param mcp_buffer A [`numeric`] distance to buffer the mcp (Default \code{0}).
 #' Only used if \code{"mcp"} is used.
 #' @param novel Which conditions are to be masked out respectively, either the
@@ -64,7 +68,7 @@
 #' @note
 #' The method \code{"zones"} is also possible directly within [distribution()].
 #'
-#' @returns Adds extrapolation control option to a [`distribution`] object.
+#' @returns Adds extrapolation limit option to a [`distribution`] object.
 #'
 #' @references
 #' * Randin, C. F., Dirnböck, T., Dullinger, S., Zimmermann, N. E., Zappa, M., & Guisan, A. (2006).
@@ -87,23 +91,23 @@
 #'  # To add a zone layer for extrapolation constraints.
 #'  x <- distribution(background) |>
 #'    add_predictors(covariates) |>
-#'    add_control_extrapolation(method = "zones", layer = zones)
+#'    add_limits_extrapolation(method = "zones", layer = zones)
 #' }
 #'
-#' @name add_control_extrapolation
+#' @name add_limits_extrapolation
 NULL
 
-#' @rdname add_control_extrapolation
+#' @rdname add_limits_extrapolation
 #' @export
 methods::setGeneric(
-  "add_control_extrapolation",
+  "add_limits_extrapolation",
   signature = methods::signature("x"),
   function(x, layer, method = "mcp", mcp_buffer = 0,
-           novel = "within", limits_clip = FALSE) standardGeneric("add_control_extrapolation"))
+           novel = "within", limits_clip = FALSE) standardGeneric("add_limits_extrapolation"))
 
-#' @rdname add_control_extrapolation
+#' @rdname add_limits_extrapolation
 methods::setMethod(
-  "add_control_extrapolation",
+  "add_limits_extrapolation",
   methods::signature(x = "BiodiversityDistribution"),
   function(x, layer, method = "mcp", mcp_buffer = 0, novel = "within", limits_clip = FALSE) {
     assertthat::assert_that(inherits(x, "BiodiversityDistribution"),
@@ -178,6 +182,50 @@ methods::setMethod(
     }
 
     # Return the altered object
+    return(y)
+  }
+)
+
+#' Remove limits from an existing distribution object
+#'
+#' @description This function allows to remove set limits from an existing [distribution]
+#' object.
+#'
+#' @param x [distribution] (i.e. [`BiodiversityDistribution-class`]) object.
+#'
+#' @family control
+#' @seealso [add_limits_extrapolation()]
+#'
+#' @examples
+#' \dontrun{
+#'  x <- distribution(background) |>
+#'    add_predictors(covariates) |>
+#'    add_limits_extrapolation(method = "zones", layer = zones)
+#'  x <- x |> rm_limits()
+#'  x
+#' }
+#'
+#' @name rm_limits
+NULL
+
+#' @rdname rm_limits
+#' @export
+methods::setGeneric(
+  "rm_limits",
+  signature = methods::signature("x"),
+  function(x) standardGeneric("rm_limits"))
+
+#' @rdname rm_limits
+methods::setMethod(
+  "rm_limits",
+  methods::signature(x = "BiodiversityDistribution"),
+  function(x) {
+    assertthat::assert_that(inherits(x, "BiodiversityDistribution") )
+    # Make a deep copy
+    y <- x$clone(deep = TRUE)
+
+    y <- y$rm_limits()
+    # Return x without limits
     return(y)
   }
 )