+++ tests/testthat/test_rangesOffsets.R
@@ -24,9 +24,11 @@ test_that('Load ranges and add them to distribution object', {
   # Now set them one up step by step
   x <- distribution(background)
   # This will raise a warning since projection is different
-  suppressMessages(
-        x |> add_predictor_range(virtual_range, method = 'distance')
+  expect_warning(
+    suppressMessages(
+      x |> add_predictor_range(virtual_range, method = 'distance')
     )
+  )
 
   # Try and add a range as raster
   virtual_range_ras <- terra::rasterize(virtual_range, background)